sweet alert toast added for add, update and delete toy operation

+++ src/Pages/MyToys/MyToys.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../AuthProviders/AuthProviders";
 import MyToysRow from "./MyToysRow";
 import UpdateToyModal from "./UpdateToyModal";
+import Swal from "sweetalert2";
 
 
 const MyToys = () => {
@@ -31,18 +32,39 @@ const MyToys = () => {
     }
 
     const handleDelete = id => {
-        fetch(`http://localhost:5000/toys/${id}`, {
-            method: 'DELETE'
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/toys/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        // console.log(data)
+                        if (data.deletedCount > 0) {
+                            
+                            const remaining = myToys.filter(toy => toy._id !== id);
+                            setMyToys(remaining);
+                            Swal.fire(
+                                'Deleted!',
+                                'Your toy has been deleted.',
+                                'success'
+                            )
+                        }
+                    })
+                
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                // console.log(data)
-                if (data.deletedCount > 0) {
-                    alert("Toy Removed!")
-                    const remaining = myToys.filter(toy => toy._id !== id);
-                    setMyToys(remaining);
-                }
-            })
+
+
     }
     return (
         <div>
